added tests and /api/v1/user-statistic

+++ src/main/java/com/running/tracker/data/response/StatisticsResponse.java
@@ -12,6 +12,6 @@ import lombok.NoArgsConstructor;
 public class StatisticsResponse {
 
     private int runsCount;
-    private int totalDistance;
-    private int avgSpeed;
+    private double totalDistance;
+    private double avgSpeed;
 }
